Resolve merge suggestions.
