+++ core/src/main/java/knaufdan/android/core/arch/IBaseActivity.kt
@@ -2,9 +2,9 @@ package knaufdan.android.core.arch
 
 import androidx.fragment.app.FragmentManager
 import knaufdan.android.core.arch.implementation.BaseViewModel
-import knaufdan.android.core.databinding.Bindable
+import knaufdan.android.core.databinding.BindableElement
 
-interface IBaseActivity<ViewModel : BaseViewModel> : Bindable<ViewModel> {
+interface IBaseActivity<ViewModel : BaseViewModel> : BindableElement<ViewModel> {
     fun getInitialPage(): Int = -1
 
     fun getTitleRes(): Int = -1
