+++ core/src/main/java/knaufdan/android/core/arch/IBaseFragment.kt
@@ -1,9 +1,9 @@
 package knaufdan.android.core.arch
 
 import knaufdan.android.core.arch.implementation.BaseViewModel
-import knaufdan.android.core.databinding.Bindable
+import knaufdan.android.core.databinding.BindableElement
 
-interface IBaseFragment<ViewModel : BaseViewModel> : Bindable<ViewModel> {
+interface IBaseFragment<ViewModel : BaseViewModel> : BindableElement<ViewModel> {
     fun getTitleRes(): Int = -1
 
     fun setBackPressed(isBackPressed: Boolean)
