--- core/src/main/java/knaufdan/android/core/databinding/Bindable.kt
@@ -1,12 +1,0 @@
-package knaufdan.android.core.databinding
-
-typealias LayoutRes = Int
-typealias BindingKey = Int
-
-interface Bindable<DataSource> {
-    fun getDataSource(): DataSource
-
-    fun getLayoutRes(): LayoutRes
-
-    fun getBindingKey(): BindingKey
-}
